refactor: rename libs/dsl → libs/lang, DSLRuntime → GaiaRuntime, DSLFactorGraph → CompiledFactorGraph
Aligns code naming with "Gaia Language" terminology used in docs and specs.
Part of Issue #82 P0.
