+++ tests/cli/test_review_store.py
@@ -10,8 +10,8 @@ from cli.review_store import (
     read_review,
     write_review,
 )
-from libs.dsl.loader import load_package
-from libs.dsl.resolver import resolve_refs
+from libs.lang.loader import load_package
+from libs.lang.resolver import resolve_refs
 
 
 FIXTURE_PATH = Path("tests/fixtures/dsl_packages/galileo_falling_bodies")
